Se crearon conexiones y se consumen entre los microservicios consultas y pacientes

+++ consultas/src/main/java/com/minsait/consultas/model/entity/HistorialMedico.java
@@ -18,7 +18,7 @@ public class HistorialMedico {
     @Column(name = "id_historial")
     private Long id;
 
-    private int idPaciente;
+    private Long idPaciente;
     @OneToMany(cascade={CascadeType.PERSIST, CascadeType.MERGE},mappedBy = "historial")
     @JsonIgnore
     private List<Consulta> consulta;
